requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Cython==0.29.7
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
ftfy==4.4.3
html5lib==1.0.1
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
lxml==4.3.3
Markdown==3.1
mccabe==0.6.1
multidict==4.5.2
numpy==1.16.3
PyJWT==1.7.1
pylint==2.3.1
PySocks==1.6.8
pytz==2019.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
twilio==6.26.3
typed-ast==1.3.4
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.1
yarl==1.3.0
